// src/app/store/actionTypes/category-action.types.ts
export enum CategoryActionTypes {
    GET_CATEGORIES = '[Categories] get categories',
    GET_CATEGORIES_OK = '[Categories] get categories ok',
    GET_CATEGORIES_FAIL = '[Categories] get categories fail',

    GET_CATEGORY = '[Categories] get category',
    GET_CATEGORY_OK = '[Categories] get category ok',
    GET_CATEGORY_FAIL = '[Categories] get category fail',

    UPDATE_CATEGORY = '[Categories] update category',
    UPDATE_CATEGORY_OK = '[Categories] update category ok',
    UPDATE_CATEGORY_FAIL = '[Categories] update category fail',

    CLEAR_CATEGORIES_ERROR = '[Categories] clear error',

    DELETE_CATEGORY = '[Categories] delete category',
    DELETE_CATEGORY_OK = '[Categories] delete category ok',
    DELETE_CATEGORY_FAIL = '[Categories] delete category fail',

    CREATE_CATEGORY = '[Categories] create category',
    CREATE_CATEGORY_OK = '[Categories] create category ok',
    CREATE_CATEGORY_FAIL = '[Categories] create category fail',

    CLEAR_CURRENT_CATEGORY = '[Categories] clear current category',
}